// frontend/src/lib/api/orval/api/generated/model/sectionList.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Langu-AI API
 * API for the Langu-AI German language learning platform
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryEnum } from './categoryEnum';
import type { ContentTypeEnum } from './contentTypeEnum';

export interface SectionList {
  readonly id: string;
  readonly level_code: string;
  category: CategoryEnum;
  /** @maxLength 255 */
  title: string;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   */
  order?: number;
  content_type: ContentTypeEnum;
  readonly item_count: number;
}